requirements.txt
aiogram==3.4.1
motor==3.6.0
Pillow==10.1.0
python-dotenv==1.0.0
